pyproject.toml
[project]
name = "img2ico"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.14.1",
]
